requirements.worker.txt
pyyaml>=6.0
kubernetes>=28.1.0
jsonschema>=4.17.3